Conditionally formats a worksheet based on data within the dataframe as a side effect.

Conditionally formats a worksheet based on data within the dataframe as a side effect.

fmt_luminex_results(
  excel_file,
  dFrame,
  sheet_name = "final_result",
  low_positive_controls_df
)

Arguments

excel_file
character vector with the name of the Excel file

dFrame
dataframe being used to guide highlighting. It is the same one as was used to make the worksheet.

sheet_name
name of the worksheet to highlight.

low_positive_controls_df
dataframe with the low positive controls

Value

Excel file name of file that has been formatted to highlight non-negative results in the sheet indicated. This highlights "I" and "P" results