Returns positive antigens based on low control value and a multiplier

From e-mail 2015-09-22 13:32 by Luis Giavedoni

get_positive_antigens(d_mfi_df, mult = 1)

Arguments

d_mfi_df
dataframe wih standardized antigen reactivity measures

mult
multiplier used to adjust cutoff level.

Details

Mark,

When the CNT LOW is below 2.5, we use the CNT LOW score to assign positive reactivity. Any score equal or higher than the CNT LOW will be scored as positive.

When the CNT LOW is above 2.5, we use 2.5 as the cut off. Any score equal or higher than 2.5 will be scored as positive.

I know that I said that the low score was 3.0, but after review of current Charles River documents, the recommendation is for the score to be 2.5.

Regards,

Luis

Thus, the function used is df$result <- ifelse(df$val >= min(2.5, df$cutoff), "P", "N")

SPF Colony Testing Plate 96 06-09-16.xlsx has an extra low control ICL CNT LOW (dup). Controls with stri_detect_fixed(tolower(name), "(dup)") == TRUE are ignored.