Form an Excel workbook and worksheet similar in format to those being made by Luis Giavedoni's laboratory.

The Excel file ``Report plate 62 08-15-2014.xlsx'' was used as a guide.

make_excel_wkbk(
  file,
  w_raw_mfi_df,
  w_mfi_df,
  w_d_mfi_df,
  w_r_mfi_df,
  w_combined_df,
  low_positive_controls_df
)

Arguments

file
character vector of length one with name of file to create

w_raw_mfi_df
dataframe with raw results from Luminex machine

w_mfi_df
dataframe with background substracted from raw results

w_d_mfi_df
dataframe with normalized values where divisors were used on background subtracted results

w_r_mfi_df
dataframe where each antigen result is called as N, I, or P based on is values relative to a fraction of the low control value.

w_combined_df
dataframe with the antigen calls combined into agent (virus) calls. (See get_combined() for algorithm).

low_positive_controls_df
dataframe with sample, name, wells, and val columns that identify the well with low positive control values (<5000 for Human or Goat IgG).